refactor: cleanup nitro plugin
